% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot A Year Of Accidents Geographically on a Map of the State they Happened In}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integer value from 1 to 50 to select the state to map accidents for}

\item{year}{The year of accident data you wish to plot as an integer (2013, 2014, or 2015)}
}
\value{
This function returns a map of the state and scatter plots accidents on it geographically
}
\description{
Plot A Year Of Accidents Geographically on a Map of the State they Happened In
}
\examples{
\dontrun{
fars_map_state(5,2013)
}
}
