% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read A Year of FARS Data Into a Dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{is a string which represnts the name of a .csv file}
}
\value{
When successful this fxn returns a data frame containing data from specificed .csv
     Returns a string "file _____ does not exist" when it cannot find the file
}
\description{
Read A Year of FARS Data Into a Dataframe
}
\examples{
\dontrun{
farsTibble <- fars_read("accident_2013.csv.bz2")
}
}
