% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Aggregate a List Of Years Worth of FARS Data Into a List of Dataframes}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{}
}
\value{
this file returns a list of data frames when successful.
   Each data frame of the list contains accident data for a single year selected.
   The the date frames are filtered so they only have the month and the year of each accident
   On error the function returns a string reading "invalid year:<year>".
}
\description{
This function makes a combined dataframe of the selected years given
   using lapply, make_filename, fars_Read, and daplyr. It adds a new column for each year.
   It has built in error checking to check that all the years selected are valid.
   If an invalid year is selected it will return "invalid year:<year>".
}
\examples{
\dontrun{
yrsAccdnts <- fars_read_years(c(2013:2015))
}
}
