% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize The Total Number of Accidents by Month for Each Year Selected}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{is a vector of years expressed as numerical values}
}
\value{
This function returns a data table of the form MONTH year1, ...YearN. This table
    has data summarizingn the total number of accidents for each year by MONTH.
}
\description{
This function collects that data for the selected years and makes a summary table.
    The summary shows the number of accidents for each year by month. It does ths using
    fars_read_years adn then grouping by year and month and summarizing the data.
}
\examples{
\dontrun{
smryFrsYrs <- fars_summarize_years(c(2013:2015))
}
}
