% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Make a FARS Data File Name Using Only The Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{can be a string or integer and must be 2013, 2014, or 2015}
}
\value{
this fxn returns a string that is the name of a fars data file of the specified year
      in the format "accident_<year>.csv.bz2
}
\description{
Make a FARS Data File Name Using Only The Year
}
\examples{
\dontrun{
make_filename("2014")
}
}
